+++ CHANGELOG.md
@@ -24,6 +24,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
   * Changed a type of `hideLabels` from a prop to modifier
   * Renamed `removeActive` -> `unsetActive`
   * Renamed `normalizeOptions` -> `normalizeItems`
+  * Now the `active` getter returns a Set with `multiple = true`
 
 * `gHint`:
   * Renamed: